refactor(models): rename SystemTemplate.DoIp module to DoIP per spec package (Rule 0007 B)
