+++ src/armodel/models/M2/AUTOSARTemplates/SystemTemplate/DoIP.py
