+++ src/armodel/models/__init__.py
@@ -63,7 +63,7 @@ from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.SwcInternalBehavior.
 from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.SwcInternalBehavior.ServiceMapping import *  # noqa: F403
 from armodel.models.M2.AUTOSARTemplates.SWComponentTemplate.SwcInternalBehavior.Trigger import *  # noqa: F403
 from armodel.models.M2.AUTOSARTemplates.SystemTemplate import *  # noqa: F403
-from armodel.models.M2.AUTOSARTemplates.SystemTemplate.DoIp import *  # noqa: F403
+from armodel.models.M2.AUTOSARTemplates.SystemTemplate.DoIP import *  # noqa: F403
 from armodel.models.M2.AUTOSARTemplates.SystemTemplate.DataMapping import *  # noqa: F403
 from armodel.models.M2.AUTOSARTemplates.SystemTemplate.DiagnosticConnection import *  # noqa: F403
 from armodel.models.M2.AUTOSARTemplates.SystemTemplate.ECUResourceMapping import *  # noqa: F403
